backend_python/requirements_fastapi.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
starlette==0.36.3
python-multipart==0.0.9
marshmallow==3.21.0
requests==2.31.0
numpy==1.26.4
pyproj==3.6.1
rasterio==1.3.9
websockets==12.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.4
tenacity==8.2.3 
pillow==10.4.0